requirements.txt
aiogram==3.7.0
python-dotenv
gigachat
PyPDF2
pandas
openpyxl
aiofiles
aiosqlite
apscheduler
ics
pytesseract==0.3.10
pdf2image==1.17.0
Pillow>=10.0.0